automation/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
playwright>=1.40.0
